Move merkle tree modles to database folder since it's metadata
